add test + code to select only desired counties
This commit adds a 'county_warnings' section to `config.yaml` which
contains a user defined list of counties that the user wishes to receive
warnings for. Also adds a small JSON file that maps county names to FIPS
codes.

The additional test and code use this data to filter only those warnings
whose 'region' field contains any of the counties in 'county_warnings'.

Minor change: use `pathlib.Path` objects for project root instead of a
string representing the path.