move to poetry and into package
This commit moves the project to Poetry for dependency management and packaging. Changes are either simple changes of the path or changes of the imports where previous imports were no longer valid.

Also moved the county name to FIPS code conversion into the config.py module and out of everywhere else.